// src/common/enums/colors.ts
enum Colors {
  WHITE = 'hsl(0, 0%, 100%)',
  GRAY = 'hsl(0, 0%, 98%)',
  LIGHT_GRAY = 'hsl(0, 0%, 82%)',
  DARK_GRAY = 'hsl(0, 0%, 52%)',
  DARKEN_GRAY = 'hsl(0, 0%, 30%)',
  BLUE = 'hsl(209, 23%, 22%)',
  PRIMARY_BLUE = 'hsl(207, 26%, 17%)',
  DARK_BLUE = 'hsl(200, 15%, 8%)',
}

export { Colors };